handwriting_processor/requirements.txt
python-dotenv
google-generativeai
PyMuPDF
Pillow
Jinja2